Added support for pretty-printing the XHTML output.
To get prettified output, simply call the XHtmlDocument::pretty() method
instead of XHtmlDocument::output().

Note that when pretty-printed, every element will appear on its own line in
the resulting XHTML, indented (by default) four spaces more than its parent
element. There is one exception to this rule though: PreElement.
PreElements will themselves be indented and placed on a line of their own,
but their contents will not be pretty printed.
This is because the contents of <pre> elements appear as-is on the resulting
page, which would be severely affected if they were pretty printed.